Removes dependency on nolhmann::json; adds Python test framework.
The JSON dependency is intended to speed up serialization. The
Python-native JSON serializer is implemented in C under the hood, and
should be better able to avoid casual copy-conversions of arguments and
results.

The test framework is a starting point (it does not exercise unhappy
paths or argument variations right now.)