tuber.py: Corrects (and tests) DocString and __dir__ handling.
Also binds tuber methods to the instance, not the class - this (in
principle) allows incompatible TuberObjects to be registered and used
simultaneously. This one of those "features" nobody is likely to ask for
up-front, but the existing (global) cache would have been profoundly
confusing if a conflict ever actually did arise.