Document and partially fix problems with JSON serialization.
In a perfect world, a series of calls

	1: fine
	2: returns something json.dumps can't serialize
	3: also fine

...would be returned as an array of responses:

	1: fine
	2: error
	3: error (noting that something happened in an earlier call)

However, this requires us to serialize each call separately. Because
httpserver only deals in std::string responses, we'd have to concatenate
the results separately - which can be a pretty big penalty when all we
want is some kind of multi-string buffer or callback scheme.

Instead, what we currently return for this kind of error is an object:

	{ error }

...as if we invoked a single (non-list) call. This is wrong, but the
actual fix is to move from libhttpserver to boost::beast or something a
little less unusual.