Attempt to clean up asyncio.ClientSession objects
The global cache (with WeakKeyDictionary) is not sufficient for ipython,
for some reason. This change is a little simplistic (for example, we do
see closed event loops and are hence unable to trigger close() calls on
captive ClientSessions).

arahlin's more complete option can be found in pone-software/perseus,
and is a little longer, a little more intrusive, but doesn't cut
corners. We may end up adopting it if this doesn't survive a little
testing.