Child context attributes
Attributes of Context objects that are themselves TuberObject attributes of the
input object now resolve as Context objects whose call interface remains that of
the parent Context.  This allows communicating with multiple registry objects in
a single Context block.  For example, the following will call methods of two
different registry entries (attr1 and attr2) on the remote with a single
network interaction:

    async with obj.tuber_context() as ctx:
        ctx.attr1.function1()
        ctx.attr2.function2()
        await ctx()