Simple tuber client (#22)
Implement a SimpleTuberClient class, with a corresponding SimpleContext.  The
API is similar to that of the asynchronous TuberClient, but all methods are
called serially.  The context object still enables pooling multiple method calls
into a single HTTP request as usual.

The asynchronous client object is now a subclass of the simple class, with the
asynchronous methods (e.g. tuber_resolve) overloaded appropriately.  Subclassing
either client object is simplified.