Fix __getattr__ for registry-level context object (#24)
As written, the Context for a TuberObject with objname=None (i.e. the server
registry) will raise an error on every __getattr__ call, because the registry
itself does not have any methods.

This PR instead sets attributes of this special context that behave like registry
entry objects, such that methods of these objects use the parent context to
submit requests.

Also add checks for resolved objects to avoid unnecessary network calls for
attributes that are not present on the server.