Optionally allow errors in a sequence of remote commands (#25)
Optionally allow errors in a sequence of remote commands.

On the server side, a "X-Tuber-Options: continue-on-error" flag causes exceptions in a list of commands not to return early. (That is, all function calls after the error occurs continue to execute to completion.)

On the client side, this "continue_on_error" flag is exposed as a Context / SimpleContext argument.

With a synchronous context (SimpleContext), a list of results is returned. Errors are mixed with successful results by returning Exceptions in places where errors occurred.

With an asyncio context (Context), we're returning Futures anyways and exceptions are raised for failing calls.