Dynamic object attributes
This PR extends the tuber API to enable dynamic attributes of TuberObjects.

Registry objects on the server may have attributes that are themselves registry
objects.  To mark an attribute as such on the server, add the `__tuber_object__`
attribute set to `True`.

Registry objects may also be containers -- lists or dicts of registry objects.
Importantly, all of the objects in a container must have the same set of
attributes.  To create a container object on the server, wrap it in a
`TuberContainer` constructor.  Containers may also be attributes of other
registry objects, so nested containers (e.g. lists of dicts) are possible.

Once a TuberObject is resolved on the client side, its attributes will mimic
those of the corresponding registry object on the server side.  Dynamic
attributes on the client object behave as TuberObjects, with access to method
calls, properties, and contexts as usual.  Container attributes can also be
iterated and indexed with a limited list-like or dict-like interface.

Context objects associated with a TuberObject can also have attributes and
behave as containers, but delegate object resolution to the server.
Importantly, a context object for a container does not know its dimensions or
keys, so cannot be iterated.

The resolve interface has been overhauled into a single metadata request per
object, with child attributes populated recursively.  The classic resolve
interface remains in place on the server for backward compatibility.

See the test suite for examples of the extended API in action.