Bring serial API in line with async using requests-futures
This PR rearranges the `SimpleContext` object to use a `FuturesSession` for
dispatching HTTP requests to a thread pool, rather than waiting for them to
complete on the main thread.  This brings the API for the `SimpleContext` object
in line with that of the async `Context` object.  `TuberObject` methods are
executed serially without `await`, and context methods return a `Future` whose
`result()` method returns the corresponding result from the server.  Calling a
context directly blocks until it responds, akin to the `await ctx()` API of the
async `Context`.

To allow for asynchronous communication with several servers, the (non-blocking)
`SimpleContext.send()` method returns a `Future` object whose (blocking)
`result()` method contains the full `requests.Response` object as processed from
the server.  The `SimpleContext.receive()` method can then be used to block and
return the results of each method executed on the server.