Expose tuber.server.main() for user API
This PR allows using the main() function in the `__main__` block of user code.
For example:

```
registry = {"my_object": MyObject()}

if __name__ == "__main__":
    from tuber.server import main

    main(registry=registry)
```

The registry script then automatically provides the `tuberd` command-line
interface, with the registry pre-loaded by the user.

This PR also replaces the fragile `exec` call with an importlib interface for
loading a registry from a file path.