Dynamic object attributes (#35)
* Dynamic object attributes

This PR extends the tuber API to enable dynamic attributes of TuberObjects.

Registry objects on the server may have attributes that are themselves registry
objects.  To mark an attribute as such on the server, add the `__tuber_object__`
attribute set to `True`.

Registry objects may also be containers -- lists or dicts of registry objects.
Importantly, all of the objects in a container must have the same set of
attributes.  To create a container object on the server, wrap it in a
`TuberContainer` constructor.  Containers may also be attributes of other
registry objects, so nested containers (e.g. lists of dicts) are possible.

Once a TuberObject is resolved on the client side, its attributes will mimic
those of the corresponding registry object on the server side.  Dynamic
attributes on the client object behave as TuberObjects, with access to method
calls, properties, and contexts as usual.  Container attributes can also be
iterated and indexed with a limited list-like or dict-like interface.

Context objects associated with a TuberObject can also have attributes and
behave as containers, but delegate object resolution to the server.
Importantly, a context object for a container does not know its dimensions or
keys, so cannot be iterated.

The resolve interface has been overhauled into a single metadata request per
object, with child attributes populated recursively.  The classic resolve
interface remains in place on the server for backward compatibility.

See the test suite for examples of the extended API in action.

* Container attribute access

A natural extension of this new API allows attribute evaluation in aggregate on
container items.  For example, one can get the value of a property for all items
in a container attribute of a tuber object:

```
>>> obj.list_container.tuber_get("property")
[1, 2, 3]
```
or
```
>>> obj.dict_container.tuber_get("property")
{'a': 1, 'b': 2, 'c': 3}
```

Or alternatively, evaluate the same method with the same arguments for all items:

```
>>> obj.list_container.tuber_call("method", arg1, arg2="test")
[1, 2, 3]
```

* Cleanup container ABI, allow non-string dict keys

* move tuber_call method to server, generalize container API

Server-side resolution will process method/property/object attributes of any
object with the `__tuber_object__` tag.  Additionally, any object can provide a
`tuber_meta()` method to add extra metadata entries for the client to parse.
The TuberContainer implementation is one such object.

* Use "recursive" for complete metadata traversal, rather than "not simple"
* fix backward compatiblity in server describe API
* use object path traversal list instead of evaluating a string
* use functools.reduce
* CMakeList: adds pytest --doctest-modules invocation
* server.TuberRegistry: don't re-use TuberResult.
* server.py: don't import TuberResult - we no longer need it
* server.py: from __future__ import annotations
* TuberRegistry in public server API
* jsonschema verification: use module references instead of getattr on a string
* Move 'invoke' and 'describe' into tuber.server.RequestHandler
* Teach jsonschema to more deeply inspect metadata responses
* Simple TuberArray class that assumes all items are instances of the same class
* Add a `__tuber_exclude__` class attribute that allows user to exclude particular attributes from the object description.
* Ensure that dunder attributes of parent classes are also excluded automatically.
* Add check_attribute method for server API

---------

Co-authored-by: Graeme Smecher <gsmecher@t0.technology>