Workflow for creating a client-only package
This PR adds a separate project for building just the client side of the tuber
interface, which has minimal dependencies and does not include any of the
compiled libraries to run the server.

The dependencies for the main tuberd library are also split into those
necessary to run the package, and those for testing or optional
features.