Drop --verbose option from server
This option currently only affects outputs to stderr from the file service
endpoint, and is largely redundant with what the file server would return to the
user anyway.  It makes some sense to remove this option so that users don't
expect verbose output from the tuber endpoint by enabling it.