Fixup for cross-version incompatibility between v0.15 server and v0.16+ client
Our metadata description changed between these versions, allowing a
single client/server call to fully describe objects in the server.
(Before that, you'd get a list of methods and properties but would have
to follow that up with a request for the values or descriptions of each
of them.)

Unfortunately, old servers don't know how to return the new metadata
format, and new clients weren't expecting the old format. This patch
teaches new client code to gracefully manage either.

Once the dust has settled a little bit and hardware is no longer
deployed with this combination of incompatible server versions, I would
like to remove the workaround in favour of a shorter codebase. Longer
term, I would also like to teach github to run cross-version tests so
this kind of incompatibility issue is automatically detected.