Cancel all futures on serialization error (#56)
If an error is raised by the server while constructing the response message,
then the response futures are not populated with either exceptions or results.
Calling .result() on such futures may cause the code to hang indefinitely.
Prevent this behavior by canceling the futures, so that any attempt to access
a result will immediately raise an error.