Fix method signature resolution
Signatures of bound methods drop the first argument ("self").  When such
signatures are dispatched to the client for resolution, the reconstructed
signature also drops the first argument in the list, under the assumption that
the signature is for an unbound method.

Fix this by dispatching the unbound method signature instead if possible, and
otherwise modifying the signature to include "self" as the first argument.